Transfer moving linechart to examples; remove old demo app
