+++ app/examples/custom/dynamic-linechart/index.jade
